added JSON test, and moved some integration tests to `tests` module
