+++ samples/security_event_1.xml
@@ -3,7 +3,7 @@
     <System>
         <Provider Name="Microsoft-Windows-Security-Auditing" Guid="54849625-5478-4994-A5BA-3E3B0328C30D">
         </Provider>
-        <EventID Qualifiers="">4608</EventID>
+        <EventID>4608</EventID>
         <Version>0</Version>
         <Level>0</Level>
         <Task>12288</Task>
@@ -12,13 +12,13 @@
         <TimeCreated SystemTime="2016-07-08 18:12:51.681640 UTC">
         </TimeCreated>
         <EventRecordID>1</EventRecordID>
-        <Correlation ActivityID="" RelatedActivityID="">
+        <Correlation>
         </Correlation>
         <Execution ProcessID="456" ThreadID="460">
         </Execution>
         <Channel>Security</Channel>
         <Computer>37L4247F27-25</Computer>
-        <Security UserID="">
+        <Security>
         </Security>
     </System>
     <EventData>
