+++ src/tests/mod.rs
@@ -1,1 +1,3 @@
 mod skeptic;
+mod test_full_samples;
+mod test_record_samples;
